requirements.txt
Django==3.2.13
djangorestframework==3.13.1
django-extensions==3.1.5
# django-rest-swagger==2.2.0
drf_yasg==1.20.0
django-model-utils==4.2.0
psycopg2==2.9.3